Refactor interface. #66
